YouTube_Transcript_MCP/requirements.txt
pyyaml
ascii_colors
fastmcp
youtube-transcript-api
pipmaster